big chunks, fix everywhere

+++ src/modules/users/useCases/deleteUser/DeleteUserUseCase.ts
@@ -1,5 +1,6 @@
 import { inject, injectable } from "tsyringe";
 
+import { AppError } from "../../../../shared/errors/AppError";
 import { IUsersRepository } from "../../repositories/IUsersRepository";
 
 interface IRequest {
@@ -14,6 +15,12 @@ class DeleteUserUseCase {
   ) {}
 
   async execute({ email }: IRequest): Promise<void> {
+    const user = await this.usersRepository.findByEmail(email);
+
+    if (!user) {
+      throw new AppError("User doesn't exists!");
+    }
+
     await this.usersRepository.delete(email);
   }
 }
